+++ aiohttp/client_reqrep.py
@@ -18,8 +18,10 @@ import aiohttp
 from . import hdrs, helpers, streams
 from .log import client_logger
 from .streams import EOF_MARKER, FlowControlStreamReader
-from .multidict import CIMultiDictProxy, MultiDictProxy, MultiDict, CIMultiDict
+from .multidict import (CIMultiDictProxy, MultiDictProxy, MultiDict,
+                        CIMultiDict)
 from .multipart import MultipartWriter
+from .protocol import HttpMessage
 
 PY_341 = sys.version_info >= (3, 4, 1)
 
@@ -39,6 +41,8 @@ class ClientRequest:
         hdrs.ACCEPT_ENCODING: 'gzip, deflate',
     }
 
+    SERVER_SOFTWARE = HttpMessage.SERVER_SOFTWARE
+
     body = b''
     auth = None
     response = None
@@ -53,7 +57,8 @@ class ClientRequest:
     # Until writer has finished finalizer will not be called.
 
     def __init__(self, method, url, *,
-                 params=None, headers=None, data=None, cookies=None,
+                 params=None, headers=None, skip_auto_headers=frozenset(),
+                 data=None, cookies=None,
                  files=None, auth=None, encoding='utf-8',
                  version=aiohttp.HttpVersion11, compress=None,
                  chunked=None, expect100=False,
@@ -77,6 +82,7 @@ class ClientRequest:
         self.update_host(url)
         self.update_path(params)
         self.update_headers(headers)
+        self.update_auto_headers(skip_auto_headers)
         self.update_cookies(cookies)
         self.update_content_encoding()
         self.update_auth(auth)
@@ -129,7 +135,7 @@ class ClientRequest:
         self.netloc = netloc
 
         scheme = url_parsed.scheme
-        self.ssl = scheme == 'https'
+        self.ssl = scheme in ('https', 'wss')
 
         # set port number if it isn't already set
         if not port:
@@ -174,8 +180,8 @@ class ClientRequest:
             else:
                 query = params
 
-        self.path = urllib.parse.urlunsplit(
-            ('', '', urllib.parse.quote(path, safe='/%:'), query, fragment))
+        self.path = urllib.parse.urlunsplit(('', '', helpers.requote_uri(path),
+                                             query, fragment))
         self.url = urllib.parse.urlunsplit(
             (scheme, netloc, self.path, '', ''))
 
@@ -191,13 +197,20 @@ class ClientRequest:
             for key, value in headers:
                 self.headers.add(key, value)
 
+    def update_auto_headers(self, skip_auto_headers):
+        self.skip_auto_headers = skip_auto_headers
+        used_headers = set(self.headers) | skip_auto_headers
+
         for hdr, val in self.DEFAULT_HEADERS.items():
-            if hdr not in self.headers:
-                self.headers[hdr] = val
+            if hdr not in used_headers:
+                self.headers.add(hdr, val)
 
         # add host
-        if hdrs.HOST not in self.headers:
+        if hdrs.HOST not in used_headers:
             self.headers[hdrs.HOST] = self.netloc
+
+        if hdrs.USER_AGENT not in used_headers:
+            self.headers[hdrs.USER_AGENT] = self.SERVER_SOFTWARE
 
     def update_cookies(self, cookies):
         """Update request cookies header."""
@@ -445,6 +458,7 @@ class ClientRequest:
 
         # set default content-type
         if (self.method in self.POST_METHODS and
+                hdrs.CONTENT_TYPE not in self.skip_auto_headers and
                 hdrs.CONTENT_TYPE not in self.headers):
             self.headers[hdrs.CONTENT_TYPE] = 'application/octet-stream'
 
@@ -598,7 +612,10 @@ class ClientResponse:
                         'Can not load response cookies: %s', exc)
         return self
 
-    def close(self, force=False):
+    def close(self, force=True):
+        if not force:
+            warnings.warn("force parameter should be True", DeprecationWarning,
+                          stacklevel=2)
         if self._closed:
             return
 
@@ -609,29 +626,31 @@ class ClientResponse:
                 return
 
         if self._connection is not None:
-            if self.content and not self.content.at_eof():
-                force = True
-
-            if force:
-                self._connection.close()
-            else:
-                self._connection.release()
-                if self._reader is not None:
-                    self._reader.unset_parser()
-
+            self._connection.close()
             self._connection = None
-        if self._writer is not None and not self._writer.done():
-            self._writer.cancel()
-            self._writer = None
+        self._cleanup_writer()
 
     @asyncio.coroutine
     def release(self):
         try:
-            chunk = yield from self.content.readany()
-            while chunk is not EOF_MARKER or chunk:
-                chunk = yield from self.content.readany()
+            content = self.content
+            if content is not None and not content.at_eof():
+                chunk = yield from content.readany()
+                while chunk is not EOF_MARKER or chunk:
+                    chunk = yield from content.readany()
         finally:
-            self.close()
+            self._closed = True
+            if self._connection is not None:
+                self._connection.release()
+                if self._reader is not None:
+                    self._reader.unset_parser()
+                self._connection = None
+            self._cleanup_writer()
+
+    def _cleanup_writer(self):
+        if self._writer is not None and not self._writer.done():
+            self._writer.cancel()
+        self._writer = None
 
     @asyncio.coroutine
     def wait_for_close(self):
@@ -640,7 +659,7 @@ class ClientResponse:
                 yield from self._writer
             finally:
                 self._writer = None
-        self.close()
+        yield from self.release()
 
     @asyncio.coroutine
     def read(self, decode=False):
@@ -649,10 +668,10 @@ class ClientResponse:
             try:
                 self._content = yield from self.content.read()
             except:
-                self.close(True)
+                self.close()
                 raise
             else:
-                self.close()
+                yield from self.release()
 
         data = self._content