+++ aiohttp/web_reqrep.py
@@ -124,7 +124,7 @@ class Request(dict, HeadersMixin):
 
         self._secure_proxy_ssl_header = secure_proxy_ssl_header
 
-    @property
+    @reify
     def scheme(self):
         """A string representing the scheme of the request.
 
@@ -173,7 +173,8 @@ class Request(dict, HeadersMixin):
 
     @reify
     def _splitted_path(self):
-        return urlsplit(self._path_qs)
+        url = '{}://{}{}'.format(self.scheme, self.host, self._path_qs)
+        return urlsplit(url)
 
     @property
     def raw_path(self):