+++ tests/test_client_session.py
@@ -29,6 +29,7 @@ class TestClientSession(unittest.TestCase):
 
     def tearDown(self):
         self.loop.close()
+        gc.collect()
 
     def make_open_connector(self):
         conn = BaseConnector(loop=self.loop)
@@ -43,9 +44,9 @@ class TestClientSession(unittest.TestCase):
                 "h2": "header2"
             }, loop=self.loop)
         self.assertEqual(
-            set(session._default_headers),
-            set([("h1", "header1"),
-                 ("h2", "header2")]))
+            sorted(session._default_headers.items()),
+            ([("H1", "header1"),
+              ("H2", "header2")]))
         session.close()
 
     def test_init_headers_list_of_tuples(self):
@@ -55,10 +56,10 @@ class TestClientSession(unittest.TestCase):
                      ("h3", "header3")],
             loop=self.loop)
         self.assertEqual(
-            set(session._default_headers),
-            set([("h1", "header1"),
-                 ("h2", "header2"),
-                 ("h3", "header3")]))
+            session._default_headers,
+            CIMultiDict([("h1", "header1"),
+                         ("h2", "header2"),
+                         ("h3", "header3")]))
         session.close()
 
     def test_init_headers_MultiDict(self):
@@ -69,10 +70,23 @@ class TestClientSession(unittest.TestCase):
                  ("h3", "header3")]),
             loop=self.loop)
         self.assertEqual(
-            set(session._default_headers),
-            set([("h1", "header1"),
-                 ("h2", "header2"),
-                 ("h3", "header3")]))
+            session._default_headers,
+            CIMultiDict([("H1", "header1"),
+                         ("H2", "header2"),
+                         ("H3", "header3")]))
+        session.close()
+
+    def test_init_headers_list_of_tuples_with_duplicates(self):
+        session = ClientSession(
+            headers=[("h1", "header11"),
+                     ("h2", "header21"),
+                     ("h1", "header12")],
+            loop=self.loop)
+        self.assertEqual(
+            session._default_headers,
+            CIMultiDict([("H1", "header11"),
+                         ("H2", "header21"),
+                         ("H1", "header12")]))
         session.close()
 
     def test_init_cookies_with_simple_dict(self):
@@ -108,8 +122,8 @@ class TestClientSession(unittest.TestCase):
         })
         self.assertIsInstance(headers, CIMultiDict)
         self.assertEqual(headers, CIMultiDict([
-            ("h1", "h1"),
-            ("h2", "header2")
+            ("h2", "header2"),
+            ("h1", "h1")
         ]))
         session.close()
 
@@ -122,8 +136,8 @@ class TestClientSession(unittest.TestCase):
         headers = session._prepare_headers(MultiDict([("h1", "h1")]))
         self.assertIsInstance(headers, CIMultiDict)
         self.assertEqual(headers, CIMultiDict([
-            ("h1", "h1"),
-            ("h2", "header2")
+            ("h2", "header2"),
+            ("h1", "h1")
         ]))
         session.close()
 
@@ -136,13 +150,29 @@ class TestClientSession(unittest.TestCase):
         headers = session._prepare_headers([("h1", "h1")])
         self.assertIsInstance(headers, CIMultiDict)
         self.assertEqual(headers, CIMultiDict([
-            ("h1", "h1"),
-            ("h2", "header2")
+            ("h2", "header2"),
+            ("h1", "h1")
         ]))
         session.close()
 
-    def _make_one(self):
-        session = ClientSession(loop=self.loop)
+    def test_merge_headers_with_list_of_tuples_duplicated_names(self):
+        session = ClientSession(
+            headers={
+                "h1": "header1",
+                "h2": "header2"
+            }, loop=self.loop)
+        headers = session._prepare_headers([("h1", "v1"),
+                                            ("h1", "v2")])
+        self.assertIsInstance(headers, CIMultiDict)
+        self.assertEqual(headers, CIMultiDict([
+            ("H2", "header2"),
+            ("H1", "v1"),
+            ("H1", "v2"),
+        ]))
+        session.close()
+
+    def _make_one(self, **kwargs):
+        session = ClientSession(loop=self.loop, **kwargs)
         params = dict(
             headers={"Authorization": "Basic ..."},
             max_redirects=2,
@@ -358,6 +388,7 @@ class TestClientSession(unittest.TestCase):
     def test_del(self):
         conn = self.make_open_connector()
         session = ClientSession(loop=self.loop, connector=conn)
+        self.loop.set_exception_handler(lambda loop, ctx: None)
 
         with self.assertWarns(ResourceWarning):
             del session
@@ -374,6 +405,7 @@ class TestClientSession(unittest.TestCase):
         conn = self.make_open_connector()
         session = ClientSession(connector=conn)
         self.assertIs(session._loop, self.loop)
+        session.close()
 
 
 class TestCLientRequest(unittest.TestCase):
@@ -386,14 +418,21 @@ class TestCLientRequest(unittest.TestCase):
         self.protocol = mock.Mock()
 
     def tearDown(self):
+        self.connector.close()
+        self.loop.stop()
+        self.loop.run_forever()
         self.loop.close()
 
     def test_custom_req_rep(self):
         @asyncio.coroutine
         def go():
+            conn = None
+
             class CustomResponse(ClientResponse):
                 @asyncio.coroutine
                 def start(self, connection, read_until_eof=False):
+                    nonlocal conn
+                    conn = connection
                     self.status = 123
                     self.reason = 'Test OK'
                     self.headers = CIMultiDictProxy(CIMultiDict())
@@ -430,4 +469,7 @@ class TestCLientRequest(unittest.TestCase):
                                               loop=self.loop)
             self.assertIsInstance(resp, CustomResponse)
             self.assertTrue(called)
+            resp.close()
+            conn.close()
+
         self.loop.run_until_complete(go())